Added Bitcoin Cash and Bitcoin Gold balance check

+++ currencies/currencyProcessor.go
@@ -1,5 +1,5 @@
 package currencies
 
 type CurrencyProcessor interface {
-	GetBalance(address string) int
+	GetBalance(address string) int64
 }
